test(extract/fortinet/product): move CPE↔range-type check to external test via export_test
Expose the table rows through export_test.go (ProductEntries) and run the
CPE→range-type consistency check from package product_test, matching the
repo's export_test pattern instead of an internal test file.
